Add pyright type checking to CI
Introduce a typecheck dependency group (pyright, pandas-stubs,
scipy-stubs) and a corresponding CI job that runs pyright against the
package. Widen two return-type-narrowing spots (univariate_cdf and the
tetrachoric neg_log_likelihood closure) with explicit float() calls,
since scipy's stubs report a wider return union than the declared
float return type.

+++ ordinalcorr/dichotomous.py
@@ -235,7 +235,7 @@ def tetrachoric(x: npt.ArrayLike, y: npt.ArrayLike) -> float:
 
         assert np.all(probs >= 0), f"{probs=}"
 
-        return -(counts @ np.log(probs))
+        return float(-(counts @ np.log(probs)))
 
     eps = 1e-10
     result = minimize_scalar(
